NULL handling improvements
Add support for specifying NULL-ness for types with built-in mapping.
Handle Oracle [N]VARCHAR2 and SQLite FLOAT oddities using this mechanism
instead of overriding it at the schema generation level.

Also use the is_null argument that is passed to value_traits::init_image()
to indicate whether the value can be NULL.