Handle query column exporting for composite value types
A composite value is represented in query_columns as a nested struct.
Even though the query_columns template instantiation is exported, VC++
for some reason doesn't appear to also export the nested structs. To
work around this, nested structs have to have the export macro in the
declaration. But that's not it: we also have to declare the nested
structs extern, just like the outer template instantiation itself.