Implement nested id support
Now the 'id' specifier can optionally include the data member path
to the id inside the composite value. For example:

#pragma db id(first)
std::pair<int, int> p;

Note that one somewhat counter-intuitive aspect of this new feature
is that the whole member marked with id ('p' in the above example)
and not just the actual id member ('p.first' in the above example)
is treated as readonly.

Such nested id also cannot be automatically assigned (auto specifier).