Initial take on build2-based build
Only C++11 is supported with GCC 4.8 and up and VC 14 and up.

This change also introduces the ODB_THREADS_CXX11 threading model. It
is the default for the build2 build but can be disabled with a global
ODB_THREADS_NONE define (global means that it should be in effect both
when building and using the library).