Remove inclusion of <odb/database.hxx> from lazy-ptr header
Its inclusion slowed down compilation significantly.

The consequence of this change is that you will now have to
make sure <odb/database.hxx> is included in source files that
initialize persistent lazy_ptr instances. But that is probably
acceptable (and is probably already the case anyway most of the
time).